Advserial-LLMs/requirements.txt
anthropic==0.52.0
openai==1.82.0
kokoro==0.9.4
gradio==5.31.0
numpy==2.2.6
python-dotenv==1.1.0
ipykernel==6.29.5
